+++ airbyte-integrations/connectors/destination-bigquery/src/main/java/io/airbyte/integrations/destination/bigquery/BigQueryAsyncStandardFlush.java
@@ -8,7 +8,7 @@ import com.google.cloud.bigquery.BigQuery;
 import com.google.common.util.concurrent.RateLimiter;
 import io.airbyte.cdk.integrations.destination.async.function.DestinationFlushFunction;
 import io.airbyte.cdk.integrations.destination.async.model.PartialAirbyteMessage;
-import io.airbyte.integrations.destination.bigquery.uploader.AbstractBigQueryUploader;
+import io.airbyte.integrations.destination.bigquery.uploader.BigQueryDirectUploader;
 import io.airbyte.protocol.models.v0.AirbyteStreamNameNamespacePair;
 import io.airbyte.protocol.models.v0.StreamDescriptor;
 import java.util.concurrent.ConcurrentMap;
@@ -24,10 +24,10 @@ public class BigQueryAsyncStandardFlush implements DestinationFlushFunction {
   private static final RateLimiter rateLimiter = RateLimiter.create(0.07);
 
   private final BigQuery bigQuery;
-  private final Supplier<ConcurrentMap<AirbyteStreamNameNamespacePair, AbstractBigQueryUploader<?>>> uploaderMap;
+  private final Supplier<ConcurrentMap<AirbyteStreamNameNamespacePair, BigQueryDirectUploader>> uploaderMap;
 
   public BigQueryAsyncStandardFlush(final BigQuery bigQuery,
-                                    final Supplier<ConcurrentMap<AirbyteStreamNameNamespacePair, AbstractBigQueryUploader<?>>> uploaderMap) {
+                                    final Supplier<ConcurrentMap<AirbyteStreamNameNamespacePair, BigQueryDirectUploader>> uploaderMap) {
     this.bigQuery = bigQuery;
     this.uploaderMap = uploaderMap;
   }
@@ -35,7 +35,7 @@ public class BigQueryAsyncStandardFlush implements DestinationFlushFunction {
   @Override
   public void flush(final StreamDescriptor decs, final Stream<PartialAirbyteMessage> stream) throws Exception {
     rateLimiter.acquire();
-    final ConcurrentMap<AirbyteStreamNameNamespacePair, AbstractBigQueryUploader<?>> uploaderMapSupplied = uploaderMap.get();
+    final ConcurrentMap<AirbyteStreamNameNamespacePair, BigQueryDirectUploader> uploaderMapSupplied = uploaderMap.get();
     final AtomicInteger recordCount = new AtomicInteger();
     stream.forEach(aibyteMessage -> {
       try {
