+++ airbyte-integrations/connectors/destination-bigquery/src/main/java/io/airbyte/integrations/destination/bigquery/uploader/BigQueryDirectUploader.java
@@ -4,29 +4,83 @@
 
 package io.airbyte.integrations.destination.bigquery.uploader;
 
+import static io.airbyte.integrations.destination.bigquery.helpers.LoggerHelper.printHeapMemoryConsumption;
+
 import com.google.cloud.bigquery.BigQuery;
 import com.google.cloud.bigquery.JobInfo.WriteDisposition;
+import com.google.cloud.bigquery.StandardTableDefinition;
+import com.google.cloud.bigquery.Table;
 import com.google.cloud.bigquery.TableId;
-import io.airbyte.integrations.destination.bigquery.BigQueryUtils;
+import com.google.cloud.bigquery.TableInfo;
+import io.airbyte.cdk.integrations.destination.async.model.PartialAirbyteMessage;
 import io.airbyte.integrations.destination.bigquery.formatter.BigQueryRecordFormatter;
 import io.airbyte.integrations.destination.bigquery.writer.BigQueryTableWriter;
-import io.airbyte.protocol.models.v0.AirbyteMessage;
-import java.util.function.Consumer;
+import java.io.IOException;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
-public class BigQueryDirectUploader extends AbstractBigQueryUploader<BigQueryTableWriter> {
+public class BigQueryDirectUploader {
 
-  public BigQueryDirectUploader(final TableId table,
-                                final BigQueryTableWriter writer,
-                                final WriteDisposition syncMode,
-                                final BigQuery bigQuery,
-                                final BigQueryRecordFormatter recordFormatter) {
-    super(table, writer, syncMode, bigQuery, recordFormatter);
+  private static final Logger LOGGER = LoggerFactory.getLogger(BigQueryDirectUploader.class);
+
+  protected final TableId table;
+  protected final WriteDisposition syncMode;
+  protected final BigQueryTableWriter writer;
+  protected final BigQuery bigQuery;
+  protected final BigQueryRecordFormatter recordFormatter;
+
+  BigQueryDirectUploader(final TableId table,
+                         final BigQueryTableWriter writer,
+                         final WriteDisposition syncMode,
+                         final BigQuery bigQuery,
+                         final BigQueryRecordFormatter recordFormatter) {
+    this.table = table;
+    this.writer = writer;
+    this.syncMode = syncMode;
+    this.bigQuery = bigQuery;
+    this.recordFormatter = recordFormatter;
+  }
+
+  public void upload(final PartialAirbyteMessage airbyteMessage) {
+    try {
+      writer.write(recordFormatter.formatRecord(airbyteMessage));
+    } catch (final IOException | RuntimeException e) {
+      LOGGER.error("Got an error while writing message: {}", e.getMessage(), e);
+      LOGGER.error(String.format(
+          "Failed to process a message for job: %s",
+          writer.toString()));
+      printHeapMemoryConsumption();
+      throw new RuntimeException(e);
+    }
+  }
+
+  public void closeAfterPush() {
+    try {
+      this.writer.close();
+    } catch (final IOException e) {
+      throw new RuntimeException(e);
+    }
+  }
+
+  public void createRawTable() {
+    // Ensure that this table exists.
+    final Table rawTable = bigQuery.getTable(table);
+    if (rawTable == null) {
+      LOGGER.info("Creating raw table {}.", table);
+      bigQuery.create(TableInfo.newBuilder(table, StandardTableDefinition.of(recordFormatter.getBigQuerySchema())).build());
+    } else {
+      LOGGER.info("Found raw table {}.", rawTable.getTableId());
+    }
   }
 
   @Override
-  protected void uploadData(final Consumer<AirbyteMessage> outputRecordCollector, final AirbyteMessage lastStateMessage) throws Exception {
-    BigQueryUtils.waitForJobFinish(writer.getWriteChannel().getJob());
-    super.uploadData(outputRecordCollector, lastStateMessage);
+  public String toString() {
+    return "BigQueryDirectUploader{" +
+        "table=" + table.getTable() +
+        ", syncMode=" + syncMode +
+        ", writer=" + writer.getClass() +
+        ", recordFormatter=" + recordFormatter.getClass() +
+        '}';
   }
 
 }
