+++ airbyte-integrations/connectors/destination-bigquery/src/main/java/io/airbyte/integrations/destination/bigquery/writer/BigQueryTableWriter.java
@@ -4,50 +4,24 @@
 
 package io.airbyte.integrations.destination.bigquery.writer;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import com.google.cloud.bigquery.Job;
 import com.google.cloud.bigquery.TableDataWriteChannel;
 import com.google.common.base.Charsets;
 import io.airbyte.cdk.integrations.base.AirbyteExceptionHandler;
-import io.airbyte.cdk.integrations.destination.s3.writer.DestinationWriter;
-import io.airbyte.commons.json.Jsons;
-import io.airbyte.protocol.models.v0.AirbyteRecordMessage;
 import java.io.IOException;
 import java.nio.ByteBuffer;
-import java.util.UUID;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-public class BigQueryTableWriter implements DestinationWriter {
+public record BigQueryTableWriter(TableDataWriteChannel writeChannel) {
 
   private static final Logger LOGGER = LoggerFactory.getLogger(BigQueryTableWriter.class);
 
-  private final TableDataWriteChannel writeChannel;
-
-  public BigQueryTableWriter(final TableDataWriteChannel writeChannel) {
-    this.writeChannel = writeChannel;
-  }
-
-  @Override
-  public void initialize() throws IOException {}
-
-  @Override
-  public void write(final UUID id, final AirbyteRecordMessage recordMessage) {
-    throw new RuntimeException("This write method is not used!");
-  }
-
-  @Override
-  public void write(final JsonNode formattedData) throws IOException {
-    writeChannel.write(ByteBuffer.wrap((Jsons.serialize(formattedData) + "\n").getBytes(Charsets.UTF_8)));
-  }
-
-  @Override
   public void write(final String formattedData) throws IOException {
     writeChannel.write(ByteBuffer.wrap((formattedData + "\n").getBytes(Charsets.UTF_8)));
   }
 
-  @Override
-  public void close(final boolean hasFailed) throws IOException {
+  public void close() throws IOException {
     this.writeChannel.close();
     try {
       final Job job = writeChannel.getJob();
@@ -59,10 +33,6 @@ public class BigQueryTableWriter implements DestinationWriter {
     } catch (final Exception e) {
       throw new RuntimeException(e);
     }
-  }
-
-  public TableDataWriteChannel getWriteChannel() {
-    return writeChannel;
   }
 
 }
